backend/requirements.txt
# Developed by Sydney Edwards
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
bcrypt>=4.0.0
pyjwt>=2.8.0
openai>=1.0.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
httpx>=0.27.0
reportlab>=4.0.0
slowapi>=0.1.9
limits>=3.13.0
stripe>=8.0.0
